requirements.txt
psycopg2-binary
nba_api
pandas
requests
django